Replace Subsemaphore with a semaphore release manager.
The assumption of the bounded gather operations is that threads of
control hold the bounding semaphore.  That means, when the bounded
gather operations should be called from a thread of control which
holds the semaphore and when those operations need to block internally
(by calling wait or gather, say) it should release the semaphore.  I
added a "WithoutSemaphore" semaphore release manager to implement this
pattern and use it in the bounded gather operations.